AttributeVersions(): remove unnecessary cast.

+++ tidy/src/attrs.c
@@ -199,7 +199,7 @@ static uint AttributeVersions(Node* node, AttVal* attval)
         return attval->dict->versions;
 
     for (i = 0; node->tag->attrvers[i].attribute; ++i)
-        if (node->tag->attrvers[i].attribute == (uint)attval->dict->id)
+        if (node->tag->attrvers[i].attribute == attval->dict->id)
             return node->tag->attrvers[i].versions;
 
     return attval->dict->versions & VERS_ALL
